pyproject.toml
[project]
name = "function-calling-ai-agent-self-project-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.5.0",
    "numpy>=2.4.2",
    "openai>=2.21.0",
    "pandas<3.0.0",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.8.0",
    "streamlit>=1.42.0",
    "tavily-python>=0.7.21",
]
